third_party/WebKit/LayoutTests/fast/dom/shadow/disconnected-distribution-cycle-expected.txt

CONSOLE WARNING: line 15: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Distributions from disconnected subtrees should be cleared when inserted again to avoid cycles.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS content.getDistributedNodes().length is 0
PASS successfullyParsed is true

TEST COMPLETE
